requirements.txt
discord.py==2.3.2
psycopg2-binary==2.9.9
PyYAML==6.0.1
pytz==2023.3
thefuzz[speedup]==0.20.0